+++ packages/cloud-cognitive/src/components/ImportModal/ImportModal.test.js
@@ -9,8 +9,10 @@ import { fireEvent, render, waitFor, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import React from 'react';
 
-import { carbon } from '../../settings';
+import { carbon, pkg } from '../../settings';
 import { ImportModal } from '.';
+
+const { devtoolsAttribute, getDevtoolsId } = pkg;
 
 global.fetch = jest.fn(() =>
   Promise.resolve({
@@ -320,14 +322,25 @@ describe(componentName, () => {
     expect(container.firstChild).toHaveClass(defaultProps.className);
   });
 
+  const dataTestId = 'data-testid';
+
   it('adds additional properties to the containing node', () => {
-    render(<ImportModal {...defaultProps} data-testid="test-id" />);
-    screen.getByTestId('test-id');
+    render(<ImportModal {...defaultProps} data-testid={dataTestId} />);
+    screen.getByTestId(dataTestId);
   });
 
   it('forwards a ref to an appropriate node', () => {
     const ref = React.createRef();
     render(<ImportModal {...defaultProps} ref={ref} />);
     expect(ref.current).not.toBeNull();
+  });
+
+  it('adds the Devtools attribute to the containing node', () => {
+    render(<ImportModal {...defaultProps} data-testid={dataTestId} />);
+
+    expect(screen.getByTestId(dataTestId)).toHaveAttribute(
+      devtoolsAttribute,
+      getDevtoolsId(componentName)
+    );
   });
 });