+++ packages/cloud-cognitive/src/components/Saving/Saving.test.js
@@ -9,7 +9,11 @@ import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import React from 'react';
 
+import { pkg } from '../../settings';
+
 import { Saving } from '.';
+
+const { devtoolsAttribute, getDevtoolsId } = pkg;
 
 const componentName = Saving.displayName;
 const defaultProps = {
@@ -82,14 +86,25 @@ describe(componentName, () => {
     expect(container.firstChild).toHaveClass(defaultProps.className);
   });
 
+  const dataTestId = 'data-testid';
+
   it('adds additional properties to the containing node', () => {
-    render(<Saving {...defaultProps} data-testid="test-id" />);
-    screen.getByTestId('test-id');
+    render(<Saving {...defaultProps} data-testid={dataTestId} />);
+    screen.getByTestId(dataTestId);
   });
 
   it('forwards a ref to an appropriate node', () => {
     const ref = React.createRef();
     render(<Saving {...defaultProps} ref={ref} />);
     expect(ref.current).not.toBeNull();
+  });
+
+  it('adds the Devtools attribute to the containing node', () => {
+    render(<Saving {...defaultProps} data-testid={dataTestId} />);
+
+    expect(screen.getByTestId(dataTestId)).toHaveAttribute(
+      devtoolsAttribute,
+      getDevtoolsId(componentName)
+    );
   });
 });