+++ packages/cloud-cognitive/src/components/SidePanel/SidePanel.test.js
@@ -16,8 +16,10 @@ import uuidv4 from '../../global/js/utils/uuidv4';
 import { SidePanel } from '.';
 import { Add16 } from '@carbon/icons-react';
 
-const blockClass = `${pkg.prefix}--side-panel`;
-const actionSetBlockClass = `${pkg.prefix}--action-set`;
+const { devtoolsAttribute, getDevtoolsId, prefix } = pkg;
+
+const blockClass = `${prefix}--side-panel`;
+const actionSetBlockClass = `${prefix}--action-set`;
 const sizes = ['xs', 'sm', 'md', 'lg', 'max'];
 
 const dataTestId = uuidv4();
@@ -28,6 +30,7 @@ const pageContentSelectorValue = '#side-panel-test-page-content';
 
 const onRequestCloseFn = jest.fn();
 const onUnmountFn = jest.fn();
+
 const renderSidePanel = ({ ...rest }, children = <p>test</p>) =>
   render(
     <SidePanel
@@ -386,6 +389,15 @@ describe('SidePanel', () => {
     const ref = React.createRef();
     renderSidePanel({ ref });
     expect(ref.current).toEqual(screen.getByRole('complementary'));
+  });
+
+  it('adds the Devtools attribute to the containing node', () => {
+    renderSidePanel({ 'data-testid': dataTestId });
+
+    expect(screen.getByTestId(dataTestId)).toHaveAttribute(
+      devtoolsAttribute,
+      getDevtoolsId(SidePanel.displayName)
+    );
   });
 
   it('should call the onRequestClose event handler', () => {